requirements.txt
django
django-crispy-forms==1.14.0
psycopg2-binary==2.9.5
dj-database-url==1.2.0
gunicorn==20.1.0
whitenoise==6.2.0
Brotli==1.0.9
Pillow
python-dotenv
django-allauth